+++ astropy/units/quantity.py
@@ -18,13 +18,12 @@ import numpy as np
 # AstroPy
 from ..extern import six
 from .core import (Unit, dimensionless_unscaled, UnitBase, UnitsError,
-                   get_current_unit_registry)
+                   get_current_unit_registry, UnitConversionError)
 from .format.latex import Latex
 from ..utils.compat import NUMPY_LT_1_7, NUMPY_LT_1_8, NUMPY_LT_1_9
 from ..utils.compat.misc import override__dir__
 from ..utils.misc import isiterable, InheritDocstrings
 from ..utils.data_info import ParentDtypeInfo
-from .utils import validate_power
 from .. import config as _config
 
 
@@ -357,17 +356,27 @@ class Quantity(np.ndarray):
         # amount that depends on one of the input values, so we need to treat
         # this as a special case.
         # TODO: find a better way to deal with this case
-        if function is np.power and result_unit is not dimensionless_unscaled:
+        if function is np.power and result_unit != dimensionless_unscaled:
 
             if units[1] is None:
                 p = args[1]
             else:
                 p = args[1].to(dimensionless_unscaled).value
 
-            result_unit = result_unit ** validate_power(p)
+            try:
+                result_unit = result_unit ** p
+            except ValueError as exc:
+                # Changing the unit does not work for, e.g., array-shaped
+                # power, but this is OK if we're (scaled) dimensionless.
+                try:
+                    converters[0] = units[0]._get_converter(
+                        dimensionless_unscaled)
+                except UnitConversionError:
+                    raise exc
+                else:
+                    result_unit = dimensionless_unscaled
 
         # We now prepare the output object
-
         if self is obj:
 
             # this happens if the output object is self, which happens
@@ -495,10 +504,20 @@ class Quantity(np.ndarray):
                 # For output arrays that require scaling, we can reuse the
                 # output array to perform the scaling in place, as long as the
                 # array is not integral. Here, we set the obj_array to `None`
-                # when it can not be used to store the scaled result.
-                if not (result_unit is None or
-                        np.can_cast(np.result_type(*inputs), obj_array.dtype)):
+                # when it cannot be used to store the scaled result.
+                # Use a try/except, since np.result_type can fail, which would
+                # break the wrapping #4770.
+                try:
+                    tmp_dtype = np.result_type(*inputs)
+                # Catch the appropriate exceptions: TypeError or ValueError in
+                # case the result_type raised an Exception, i.e. inputs is list
+                except (TypeError, ValueError):
                     obj_array = None
+                else:
+                    # Explicitly check if it can store the result.
+                    if not (result_unit is None or
+                            np.can_cast(tmp_dtype, obj_array.dtype)):
+                        obj_array = None
 
                 # Re-compute the output using the ufunc
                 if function.nin == 1: